Use decimals precision

+++ HungarianAlgorithm/Hungarian/VertexLabelledGraph.cs
@@ -3,32 +3,32 @@
 namespace Hungarian
 {
     // Decorator of IUndirectedGraph
-    public class VertexLabelledGraph : IUndirectedGraph<int, TaggedUndirectedEdge<int, double>>
+    public class VertexLabelledGraph : IUndirectedGraph<int, TaggedUndirectedEdge<int, decimal>>
     {
-        private readonly IUndirectedGraph<int, TaggedUndirectedEdge<int, double>> _graph;
-        private readonly double[] _labelling;
+        private readonly IUndirectedGraph<int, TaggedUndirectedEdge<int, decimal>> _graph;
+        private readonly decimal[] _labelling;
 
         /// <summary>
         /// Wraps the <paramref name="graph"/>, assigning 0 to every vertex label.
         /// </summary>
         /// <param name="graph"></param>
-        public VertexLabelledGraph(IUndirectedGraph<int, TaggedUndirectedEdge<int, double>> graph)
+        public VertexLabelledGraph(IUndirectedGraph<int, TaggedUndirectedEdge<int, decimal>> graph)
         {
             _graph = graph;
-            _labelling = new double[graph.VertexCount];
+            _labelling = new decimal[graph.VertexCount];
         }
 
-        public double GetVertexLabel(int vertex)
+        public decimal GetVertexLabel(int vertex)
         {
             return _labelling[vertex];
         }
 
-        public void SetVertexLabel(int vertex, double labelValue)
+        public void SetVertexLabel(int vertex, decimal labelValue)
         {
             _labelling[vertex] = labelValue;
         }
 
-        public void AddValueToVertexLabel(int vertex, double valueToAdd)
+        public void AddValueToVertexLabel(int vertex, decimal valueToAdd)
         {
             _labelling[vertex] += valueToAdd;
         }
@@ -39,18 +39,18 @@ namespace Hungarian
         public bool AllowParallelEdges => _graph.AllowParallelEdges;
         public bool IsEdgesEmpty => _graph.IsEdgesEmpty;
         public int EdgeCount => _graph.EdgeCount;
-        public IEnumerable<TaggedUndirectedEdge<int, double>> Edges => _graph.Edges;
+        public IEnumerable<TaggedUndirectedEdge<int, decimal>> Edges => _graph.Edges;
         public bool IsVerticesEmpty => _graph.IsVerticesEmpty;
         public int VertexCount => _graph.VertexCount;
         public IEnumerable<int> Vertices => _graph.Vertices;
 
-        public IEnumerable<TaggedUndirectedEdge<int, double>> AdjacentEdges(int vertex) => _graph.AdjacentEdges(vertex);        
+        public IEnumerable<TaggedUndirectedEdge<int, decimal>> AdjacentEdges(int vertex) => _graph.AdjacentEdges(vertex);        
         public int AdjacentDegree(int vertex) => _graph.AdjacentDegree(vertex);
         public bool IsAdjacentEdgesEmpty(int vertex) => _graph.IsAdjacentEdgesEmpty(vertex);
-        public TaggedUndirectedEdge<int, double> AdjacentEdge(int vertex, int index) => _graph.AdjacentEdge(vertex, index);
-        public bool TryGetEdge(int source, int target, out TaggedUndirectedEdge<int, double> edge) => _graph.TryGetEdge(source, target, out edge);
+        public TaggedUndirectedEdge<int, decimal> AdjacentEdge(int vertex, int index) => _graph.AdjacentEdge(vertex, index);
+        public bool TryGetEdge(int source, int target, out TaggedUndirectedEdge<int, decimal> edge) => _graph.TryGetEdge(source, target, out edge);
         public bool ContainsEdge(int source, int target) => _graph.ContainsEdge(source, target);
-        public bool ContainsEdge(TaggedUndirectedEdge<int, double> edge) => _graph.ContainsEdge(edge);
+        public bool ContainsEdge(TaggedUndirectedEdge<int, decimal> edge) => _graph.ContainsEdge(edge);
         public bool ContainsVertex(int vertex) => _graph.ContainsVertex(vertex);
         #endregion // Decorator members
     }
